backend/requirements.txt
# FastAPI Core
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1

# Database & Storage
pymongo==4.10.1
chromadb==0.5.23
sentence-transformers==3.3.1
numpy==2.1.3
pandas==2.2.3
PyMuPDF==1.24.10
python-docx==1.1.2

# AI & ML
openai==1.57.0
google-generativeai==0.8.3
pythainlp==5.1.2
torch==2.5.1
transformers==4.46.2

# Utilities
pydantic==2.10.3
pydantic-settings==2.6.1
httpx==0.28.1
aiofiles==24.1.0
Pillow==11.0.0
shortuuid==1.0.13
python-dateutil==2.9.0.post0